+++ statusBarItem.js
@@ -4,6 +4,7 @@ class statusBarItem {
 	constructor() {
 		this.str = ""
 		this.act = false
+		this.arr = new Array(100)
 		this.config = {};
         this.activate = () => {
             if (this.statusBarItem) {
@@ -23,7 +24,25 @@ class statusBarItem {
             if (!this.statusBarItem) {
                 return;
 			}
-            this.statusBarItem.text = this.str.replace("{count}", count);
+			var mod100 = count % 100
+			var speed
+			var time = (new Date()).getTime()
+			if(this.arr[mod100] != undefined) {
+				speed = 6e6 / (time - this.arr[mod100])
+				this.arr[mod100] = time
+			} else {
+				var bg = (mod100 + 99) % 100
+				while(this.arr[bg] != undefined) {
+					bg = (bg + 99) % 100
+				}
+				if(bg == (mod100 + 99) % 100) {
+					speed = .0
+				} else {
+					speed = (mod100 + 99 - bg) % 100 * 6e4 / (time - this.arr[(bg + 1) % 100])
+				}
+				this.arr[mod100] = time
+			}
+            this.statusBarItem.text = this.str.replace("{count}", count).replace("{speed}", speed.toFixed(2))
 		};
 		this.onDidChangeTextDocument = (count, event) => {
             this.updateStatusBar(count);