+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.1.0
+
++ Added speed displaying function.
++ Fixed a bug that the counter will display 0 when vscode starts.
+
 ## 1.0.0
 
 - Initial release